presenter metl ui tweaks and bug fixes
Former-commit-id: 17d6a3ffaeb74e78ec3c61f7d5f4625b99aa9332

+++ MeTLMeeting/SandRibbonInterop/LocalCache.cs
@@ -83,17 +83,45 @@ namespace SandRibbonInterop.LocalCache
         {
             if (remoteUri.ToString().StartsWith(cacheName + "\\"))
                 return remoteUri;
+            var file = remoteUri.ToString().Split('/').Reverse().First();
+            if(file.Contains("submission"))
+            {
+                int num;
+                if (!(int.TryParse(file.Substring(0, 1), out num)))
+                    return remoteUri;
+            }
             if (!CacheDict.ContainsKey(remoteUri.ToString()))
             {
                 if (!Directory.Exists(cacheName))
                     Directory.CreateDirectory(cacheName);
-                var localUriString = cacheName + "\\" + remoteUri.ToString().Split('/').Reverse().First();
-                File.WriteAllBytes(localUriString, HttpResourceProvider.secureGetData(remoteUri.ToString()));
+                var localUriString = generateLocalFilename(remoteUri.ToString());
+                //check if file exists
+                if(!File.Exists(localUriString))
+                    File.WriteAllBytes(localUriString, HttpResourceProvider.secureGetData(remoteUri.ToString()));
                 var localUri = new Uri(localUriString, UriKind.Relative);
                 Add(remoteUri.ToString(), localUri);
             }
             return CacheDict[remoteUri.ToString()];
         }
+
+        private static string generateLocalFilename(string uri)
+        {
+            var uriParts = uri.Split('/').Reverse();
+            var filename = uriParts.First();
+            int tmp;
+            if (int.TryParse(uriParts.ElementAt(1), out tmp))
+            {
+                var slideDirectory = uriParts.ElementAt(1);
+                if (!Directory.Exists(cacheName + "\\" + slideDirectory))
+                    Directory.CreateDirectory(cacheName + "\\" + slideDirectory);
+                return cacheName + "\\" + slideDirectory + "\\" + filename;
+            }
+            else
+            {
+                return cacheName + "\\" + filename;
+            }
+        }
+
         public static Uri RemoteSource(Uri media)
         {
